Add checkstyle for static code analysis
The code in general is not very well formatted and also there are some
pitfalls which could be prevented using a tool for static code analysis
like checkstyle or PMD. In this step only checkstyle has been
introduced.

Even with the fact that checkstyle is optional at the moment to not
break future builds it may help to clean up code and make it more
readable and help on refactoring. Only readable code is maintainable!

The configuration beeing added with this patch is based on Google's
Java Coding Guidelines.

http://google.github.io/styleguide/javaguide.html

This only difference is the indentation beeing set to 4 whitespaces
instead of 2.

Signed-off-by: Akeshihiro <akeshihiro@gmx.net>