Rewrite .gitignore file
The .gitignore file contained only ignore definitions primary for Maven
and NetBeans and also was not structured, so you had to interpret the
ignores.

The .gitignore file has been rewritten. The ignores are grouped by the
tools. Also the ignore definitions are handled recursively.

Groups of files, that are ignore:
- Maven (generated files)
- Eclipse (project files and generated files)
- IntelliJ (project files and generated files)
- NetBeans (project files and generated files)

Signed-off-by: Akeshihiro <akeshihiro@gmx.net>